dataplane-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
torch==2.1.1
transformers==4.36.0
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
loguru==0.7.2
motor==3.3.2
pymongo==4.6.0
pytest==7.4.3
pytest-asyncio==0.21.1 